% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.diag.R
\name{find.diag}
\alias{find.diag}
\title{Find diagnostic markers and haplotypes for known aboveground samples}
\usage{
find.diag(
  ind.list,
  stacks.dat,
  max.md.marker = 0.2,
  max.md.hap = 0.2,
  min.dep = 20,
  max.haps = 2,
  verbose = FALSE
)
}
\arguments{
\item{ind.list}{a named list with a vector for each species containing the names of all individuals of that species}

\item{stacks.dat}{output of read.stacks}

\item{max.md.marker}{the maximum proportion of missing data among individuals of the focal species to call a species-diagnostic marker}

\item{max.md.hap}{the maximum proportion of missing data in other individuals of a species to call an individual-diagnostic haplotype}

\item{min.dep}{the minimum read depth required for a marker to be considered present in an individual}

\item{max.haps}{the maximum number of haplotypes allowed per-marker (no limit if NA)}

\item{verbose}{whether to output progress information}
}
\value{
Returns a list, intended as input for match.diag, with the following elements:
\itemize{
\item{species.diag.markers:}{ a list containing a vector of diagnostic marker index numbers for each species}
\item{individual.diag.haplotypes:}{ a list of lists, one for each species, containing a vector of diagnostic haplotype index numbers for each individual of the species, and one of non-individual-diagnostic haplotypes (non.diagnostic)}
\item{summary:}{ a list containing dataframes with counts of diagnostic markers and haplotypes for each species and individual}
}
}
\description{
Finds markers which distinguish one species in the dataset from all other species (species-diagnostic markers) and haplotypes within these which distinguish one individual of the species from all other individuals (individual-diagnostic haplotypes).
}
\details{
This function takes the output of read.stacks and identifies diagnostic markers and haplotypes which can be used to identify species and individuals in mixed unknown samples. It identifies two classes of diagnostic markers: 1. Species-diagnostic markers, these are markers that are present in at least the proportion of individuals of a species set by the inverse of max.md.marker but not in any individuals outside the species. 2. Individual-diagnostic haplotypes, which are haplotypes which are only present in one individual, although the marker must be present (with a different haplotype) in other individuals of the species.
}
