% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.diag.R
\name{match.diag}
\alias{match.diag}
\title{Match unknown root samples to diagnostic markers and haplotypes}
\usage{
match.diag(data.dir, diag, min.reads.mar = 1, min.reads.hap = 1)
}
\arguments{
\item{data.dir}{the path to the directory which contains the stacks 'matches.tsv.gz' files for the unknown root samples}

\item{diag}{list containing diagnostic markers made by find.diag}

\item{min.reads.mar}{minimum read depth for species-diagnostic marker matches}

\item{min.reads.hap}{minimum read depth for individual-diagnostic haplotype matches}
}
\value{
Returns a list with the following elements:
\itemize{
\item{species.diag.markers:}{ a data frame with one row for each root sample and a column for each species containing counts of species-diagnostic markers detected in each root sample}
\item{individual.diag.haplotypes:}{ a list of data frames, one for each species, with one row for each root sample and a column for each individual, containing the proportion of detected species-diagnostic markers for the focal species which had diagnostic haplotypes for each individal}
\item{N.reads}{ a data frame containing the number of catalogue-matching reads in each sample}
}.
}
\description{
Matches markers from unknown belowground samples to diagnostic markers from aboveground samples produced by find.diag
}
\details{
Takes the diagnostic markers and haplotypes identified by find.diag and a directory with matches.tsv.gz files from STACKS for all unknown root samples. It then matches the root samples to the diagnostic markers and haplotypes. Minimum depth filters can be set with min.reads.mar and min.reads.hap.
}
