% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.stacks.R
\name{read.stacks}
\alias{read.stacks}
\title{Read stacks tsv.gz files for RootID}
\usage{
read.stacks(data.dir, min.dep = 3, verbose = F)
}
\arguments{
\item{data.dir}{path to the directory containing aboveground tsv.gz files}

\item{min.dep}{minimum depth to include haplotypes}

\item{verbose}{whether to output progress information}
}
\value{
a list containing two data frames, containing the depth of each marker (by.marker) or haplotype (by.haplotype) in each individual.
}
\description{
Read stacks *.tsv.gz files for known aboveground samples and convert them into input for the RootID package.
}
\details{
This function is the first step in the RootID package, the results is the main input for the find.diag function. All aboveground samples should be in a single directory (without belowground samples) and should all have the "matches.tsv.gz" suffix. If min.dep is set too low, it can use a large amount of memory, the default works fine on a standard laptop with our test data but larger datasets may require more memory or a higher min.dep setting.
}
