% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.rarefaction.R
\name{sample.rarefaction}
\alias{sample.rarefaction}
\title{Run rarefaction analysis on root samples}
\usage{
sample.rarefaction(
  data.dir,
  diag,
  n.subsamples = 10,
  n.reps = 10,
  verbose = TRUE
)
}
\arguments{
\item{data.dir}{he path to the directory which contains the stacks 'matches.tsv.gz' files for the unknown root samples}

\item{diag}{output of find.diag}

\item{n.subsamples}{number of subsamples}

\item{n.reps}{number of random replicates per subsample}

\item{verbose}{whether to output progress information}
}
\value{
Returns a list of data frames. n.reads is a data frame with one subsample and a column for each root sample containing the size of each subsample. Species, individual, marker and haplotype are lists, each containing dataframes with the same dimensions as n.reads with the mean, standard deviation (sd), upper (upper95CI) and lower (lower95CI) 95\% confidence intervals.
}
\description{
Subsample root data and calculate the number of markers across different subsample sizes to estimate the sufficiency of sequencing.
}
\details{
This function attempts to estimate whether the sequencing depth for each root sample is sufficient to recover all the diversity in the sample at the level of species-diagnostic markers, individual-diagnostic haplotypes, number of species and number of individuals. It uses a rarefaction-based approach similar to those used in metabarcoding studies. It randomly subsamples the data n.reps times across a range of subsample sizes (determined by n.subsamples). It then calculates the number of diagnostic haplotypes, markers, species and individuals in each subsample. The mean of these can then be plotted against subsample size. Steeper final curves indicate lower sequencing sufficiency whereas if the curve goes flat, this indicates that most diversity is likely to have been sampled.
}
