% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.marker.tree.R
\name{shared.marker.tree}
\alias{shared.marker.tree}
\title{Make dendrogram from shared markers}
\usage{
shared.marker.tree(stacks.dat, ind.list, root = NULL, collapse.to.species = F)
}
\arguments{
\item{stacks.dat}{output of read.stacks}

\item{ind.list}{a named list with a vector for each species containing the names of all individuals of that species}

\item{root}{the name of the outgroup species (or a vector of species names of the outgoup clade) by which to root the dendrogram}

\item{collapse.to.species}{if TRUE, conspecific individuals are collapsed to produce a "species tree"}
}
\value{
an object of class phylo
}
\description{
Takes the output of read.stacks and produces a rooted UPGMA tree based on shared marker presence.
}
\details{
For very closely related species, it is possible that marker presence/absence is not a reliable way to distinguish them. This function produces an Unweighted Pair Group Method with Arithmetic Mean dendrogram of all individuals based on shared marker presence, which can be plotted with e.g ape::plot.phylo. If all conspecific individuals cluster monophyletically, then it suggests that marker presence-absence does distinguish the species in the dataset. This is not intended to produce a phylogenetic tree!
}
